+++ test_browser.py
@@ -1,0 +1,15 @@
+from browser import Browser
+from url import URL
+from test_utils import socket
+
+
+class TestBrowser:
+    def test_browser_load(self):
+        _ = socket.patch().start()
+        url = "http://test.test/example1"
+        socket.respond(
+            url, b"HTTP/1.0 200 OK\r\n" + b"Header1: Value1\r\n\r\n" + b"Body text"
+        )
+        browser = Browser()
+        browser.load(URL(url))
+        assert browser.display_list is not None
